pyproject.toml
[project]
name = "face-recognition-on-edge"
version = "0.1.0"
description = "A comprehensive face recognition pipeline for edge computing"
readme = "README.md"
authors = [
    { name = "Kunal Suri", email = "kunal.suri.ml.experiments@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "insightface>=0.7.3",
    "kaggle>=1.7.4.5",
    "kagglehub>=0.3.12",
    "matplotlib>=3.10.5",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tell hatchling where to find the packages
[tool.hatchling.build.targets.wheel]
packages = ["src/face_recognition_on_edge"]
